Updates to read Deep Learning predictions
Added ability to read a time series of predictions, not just the detections.
Added empty struct fields to clicks because an exception is thrown if using multi channel data where some are grouped and some are not.
Added a channel option for loadPamGuardBinary file which filters data by channel.
Used fullfile for dirsub which means it works on MacOS and Linux